requirement.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.12.0
astor==0.8.1
bce-python-sdk==0.9.54
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
decorator==5.2.1
easyocr==1.7.2
filelock==3.20.0
fsspec==2025.10.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.7
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
modelscope==1.32.0
mpmath==1.3.0
networkx==3.6
ninja==1.13.0
numpy==2.3.5
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.10
pandas==2.2.2
pillow==12.0.0
prettytable==3.17.0
protobuf==3.20.2
psutil==7.1.3
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.1.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytesseract==0.3.13
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tifffile==2025.10.16
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
